feat: add packages and convert to ts
