+++ .eslintrc.json
@@ -1,0 +1,16 @@
+{
+  "extends": [
+    "airbnb-base",
+    "plugin:@typescript-eslint/recommended",
+    "plugin:prettier/recommended" 
+  ],
+  "parser": "@typescript-eslint/parser",
+  "plugins": [
+    "@typescript-eslint",
+    "prettier",
+    "eslint-plugin-prettier"
+  ],
+  "rules": {
+    "prettier/prettier": "error"
+  }
+}
