+++ .gitignore
@@ -1,5 +1,6 @@
 # dependencies
 /node_modules
+/dist
 
 # System Files
 .DS_Store
